Final C-Camellia version

+++ src/encryption/c/camellia/camellia_cbc.h
@@ -3,12 +3,12 @@
 
 #include "implementation.h"
 
-// Camellia-CBC encryption and decryption functions
-unsigned char* camellia_cbc_encrypt(camellia_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* camellia_cbc_decrypt(camellia_context_t* context, const unsigned char* data, int data_length, int* output_length);
+// encryption and decryption functions
+unsigned char* camellia_cbc_encrypt(camellia_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
+unsigned char* camellia_cbc_decrypt(camellia_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
-// Custom Camellia-CBC encryption and decryption functions
-unsigned char* camellia_cbc_custom_encrypt(camellia_context_t* context, const unsigned char* data, int data_length, int* output_length);
-unsigned char* camellia_cbc_custom_decrypt(camellia_context_t* context, const unsigned char* data, int data_length, int* output_length);
+// custom functions
+unsigned char* camellia_cbc_custom_encrypt(camellia_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
+unsigned char* camellia_cbc_custom_decrypt(camellia_context_t* context, const unsigned char* data, size_t data_length, size_t* output_length);
 
 #endif /* CAMELLIA_CBC_H */
